+++ extensions/positron-runtime-debug/src/util.ts
@@ -2,7 +2,9 @@
  *  Copyright (C) 2025 Posit Software, PBC. All rights reserved.
  *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
  *--------------------------------------------------------------------------------------------*/
+
 import * as vscode from 'vscode';
+import { DebugProtocol } from '@vscode/debugprotocol';
 
 export class DisposableStore implements vscode.Disposable {
 
@@ -19,5 +21,25 @@ export class DisposableStore implements vscode.Disposable {
 		}
 
 		this._disposables.clear();
+	}
+}
+
+export function formatDebugMessage(message: DebugProtocol.ProtocolMessage): string {
+	switch (message.type) {
+		case 'request': {
+			const request = message as DebugProtocol.Request;
+			return `${request.command} #${request.seq}: ${JSON.stringify(request.arguments)}`;
+		}
+		case 'event': {
+			const event = message as DebugProtocol.Event;
+			return `${event.event}: ${JSON.stringify(event.body)}`;
+		}
+		case 'response': {
+			const response = message as DebugProtocol.Response;
+			return `${response.command} #${response.request_seq}: ${JSON.stringify(response.body)}`;
+		}
+		default: {
+			return `[${message.type}]: ${JSON.stringify(message)}`;
+		}
 	}
 }
